requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pydantic
numpy
pillow
matplotlib
flask
requests